Linked to new repo
A year ago, we created Legacy Code Rocks on a whim after a bunch of us met at a software conference. Now, we have a podcast, an active slack channel, a newsletter, and more. It's time it had its own organization so we can properly manage it as an open source project. https://github.com/legacycoderocks